+++ addOns/reports/CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Maintenance changes.
 
+### Fixed
+- Include all relevant alerts in XML report templates (Issue 6627).
+
 ## [0.3.0] - 2021-05-06
 ### Added
 - API Support.